requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anthropic==0.34.2
anyio==3.7.1
appdirs==1.4.4
art==6.2
async-timeout==4.0.3
attrs==23.1.0
auto-gptq==0.5.0
autocorrect==2.6.1
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bitsandbytes==0.41.0
black==23.12.1
blinker==1.4
blis==0.7.11
boto3==1.35.19
botocore==1.35.19
bpe==1.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2023.11.17
chardet==4.0.0
charset-normalizer==3.3.1
click==8.1.7
cloud-init==24.1.3
cloudpathlib==0.16.0
cloudpickle==3.0.0
cmake==3.25.0
colorama==0.4.4
command-not-found==0.3
confection==0.1.4
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==3.4.8
cycler==0.12.1
cymem==2.0.8
datasets==2.14.6
datasketch==1.6.5
dbus-python==1.2.18
deepspeed==0.9.5
dill==0.3.7
diskcache==5.6.3
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.1.3
fastapi==0.104.0
fastchat==0.1.0
ffmpy==0.3.1
filelock==3.13.0
flash-attn==2.3.3
fonttools==4.43.1
frozenlist==1.4.0
fschat==0.2.36
fsspec==2023.10.0
gekko==1.0.6
gguf==0.9.1
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.66.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.1
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gpg==1.16.0
gradio==3.35.2
gradio_client==0.2.9
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
hjson==3.1.0
html5lib==1.1
httpcore==0.17.3
httplib2==0.20.2
httptools==0.6.1
httpx==0.24.0
huggingface-hub==0.19.4
hyperlink==21.0.0
hypothesis==6.112.1
idna==3.10
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==2.0.0
interegular==0.3.3
iotop==0.6
isort==5.13.2
jaxtyping==0.2.33
jeepney==0.7.1
Jinja2==3.1.2
jiter==0.5.0
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.19.1
jsonschema-specifications==2023.11.1
keyring==23.5.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lark==1.2.2
latex2mathml==3.77.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Levenshtein==0.25.1
linkify-it-py==2.0.2
lit==15.0.7
-e git+https://github.com/haotian-liu/LLaVA.git@f16049a3d35e281a4d88b7a2cb0a7f36a0c677ef#egg=llava
llvmlite==0.43.0
lm-format-enforcer==0.10.6
lxml==4.8.0
marisa-trie==1.2.0
markdown-it-py==2.2.0
markdown2==2.4.10
MarkupSafe==2.1.2
matplotlib==3.8.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistral_common==1.4.0
mistralai==0.4.2
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy==1.11.2
mypy-extensions==1.0.0
narwhals==1.9.1
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.0
nh3==0.2.18
ninja==1.11.1.1
nltk==3.8.1
numba==0.60.0
numpy==2.1.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
openai==1.45.0
orjson==3.10.7
outlines==0.0.46
packaging==23.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post4
pathspec==0.12.1
pathtools==0.1.2
peft==0.4.0
pexpect==4.8.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
-e git+https://github.com/vaibkumr/prompt-optimizer.git@e57a6283cda1491cd8dd00fd9e29b10735fc5806#egg=prompt_optimizer
prompt_toolkit==3.0.47
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
PuLP==2.9.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==14.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycountry==24.6.1
pycurl==7.44.1
pydantic==1.10.13
pydantic_core==2.14.3
pydub==0.25.1
Pygments==2.16.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pynvml==11.5.0
pyOpenSSL==21.0.0
pyparsing==3.1.1
pyrsistent==0.18.1
pyserial==3.5
pytest==8.3.3
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-magic==0.4.24
python-multipart==0.0.12
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.2.0
rapidfuzz==3.9.6
ray==2.35.0
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rich==13.8.0
rouge==1.0.1
rpds-py==0.10.6
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.11.3
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semantic-version==2.10.0
sentence-transformers==3.1.0
sentencepiece==0.1.99
sentry-sdk==1.32.0
service-identity==18.1.0
setproctitle==1.3.3
shapely==2.0.6
shortuuid==1.0.11
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.5.6
soupsieve==2.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
ssh-import-id==5.11
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
systemd-python==234
tenacity==9.0.0
thinc==8.2.5
threadpoolctl==3.2.0
tiktoken==0.7.0
timm==0.6.13
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
triton==2.0.0
Twisted==22.1.0
typeguard==2.13.3
typer==0.9.4
typing_extensions==4.12.2
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uc-micro-py==1.0.2
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.23.2
uvloop==0.20.0
vllm==0.6.1.post2
vllm-flash-attn==2.6.1
wadllib==1.3.6
wandb==0.18.3
wasabi==1.1.3
watchfiles==0.24.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
websockets==12.0
xformers==0.0.27.post2
xkit==0.0.0
xxhash==3.4.1
yarl==1.14.0
zipp==1.0.0
zope.interface==5.4.0
